+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Version 2.18.0
 
-### Improvements
+### Scripting
 - It is now possible to integrate with the 'QR Droid', 'Barcode Scanner' or 'Binary Eye' app to scan barcodes and use their contents in Scripting, via the new `scanBarcode()` function
+- There is now a `getLocation()` function which allows you to query your device's physical location
 
 ### Bugfixes
 - All webviews (e.g. for displaying responses) are not properly stopped when their window is closed, preventing them from using resources in the background